adding Spring-Cloud-Consul

+++ Course2/infytel-friend-family/src/main/java/com/infosys/infytel/friend/repository/FriendFamilyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.infosys.infytel.friend.entity.FriendFamily;
 
-public interface FriendFamilyRepository extends JpaRepository<FriendFamily, Integer> {
+public interface FriendFamilyRepository extends JpaRepository<FriendFamily, Long> {
 	
 	List<FriendFamily> getByPhoneNo(Long phoneNo) ;
